% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretize.R
\name{discretize}
\alias{discretize}
\title{Discretize badly}
\usage{
discretize(x, ...)
}
\arguments{
\item{x}{object decomposable to segments}

\item{...}{ignored}
}
\value{
haha nothing
}
\description{
Low memory and low performance and error-prone version of fasterize.
}
\examples{
library(raster)
library(silicate)
sc <- SC(inlandwaters)
x <- filter(sc, object_ == sample(sc$object$object_, 1))
discretize(x)
}
