% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretize.R
\name{discretize}
\alias{discretize}
\title{Discretize, not too badly}
\usage{
discretize(x, nrow = 50, ncol = 30, ..., draw = TRUE)
}
\arguments{
\item{x}{object decomposable to segments}

\item{...}{ignored}
}
\value{
haha nothing
}
\description{
Low memory and low performance and error-prone version of fasterize.
}
\examples{
library(raster)
library(silicate)
sc <- SC(inlandwaters)
x <- sc #filter(sc, object_ == sample(sc$object$object_, 3))
discretize(x, 1500, 1300)


discretize(SC0(minimal_mesh), 550, 550)
}
